Move Plots plugin to its own module
