+++ spyder/plugins/plots/widgets/figurebrowser.py
